Add base64 encoding support for bytes
According to https://protobuf.dev/programming-guides/proto3/#json
'bytes' is base64 encoded with padding.
For decoding both with/without padding and standard/url-safe encoding is
accepted.
TODO: for decoding might have to try another alphabet if standard failed

Note that this uses a 'bytes' type on the OCaml side, for backward
compatibility: 'string' would work just as well because binary values,
including \0 are valid inside OCaml strings.
This does perform extra copying now that may be unnecessary if we change the
type to be 'string' (can't use the unsafe version because we cannot guarantee
lack of mutation).

Signed-off-by: Edwin Török <edwin.torok@cloud.com>